swiftshader: llvm-16.0 changes to enable lavapipe
The LLVM situation in Android is *not good*.  There's
like 10 different llvm versions, many which are co-located
with the project that uses them (rather than a single,
stable version).

But to fix this, we probably want to deprecate Swiftshader
as one of the steps.  But we'll need to enable lavapipe
first.  So use one of the Swiftshader LLVMs now so that we
can eliminate some of the extra LLVMs long-term.

BUG=353540875
TEST=m vulkan.lvp

Change-Id: If20fde43c1b8f2da2c2e2bd0785524cfe1b0848c